+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/conf/constraint/MinMaxNormConstraint.java
@@ -53,7 +53,7 @@ public class MinMaxNormConstraint extends BaseConstraint {
      *                       be dimension 1. For CNNs, this should be dimensions [1,2,3] corresponding to last 3 of
      *                       parameters which have order [depthOut, depthIn, kH, kW]
      */
-    public MinMaxNormConstraint(double min, double max, int... dimensions){
+    public MinMaxNormConstraint(double min, double max, long... dimensions){
         this(min, max, DEFAULT_RATE, null, dimensions);
     }
 
@@ -67,7 +67,7 @@ public class MinMaxNormConstraint extends BaseConstraint {
      *                       be dimension 1. For CNNs, this should be dimensions [1,2,3] corresponding to last 3 of
      *                       parameters which have order [depthOut, depthIn, kH, kW]
      */
-    public MinMaxNormConstraint(double min, double max, double rate, int... dimensions){
+    public MinMaxNormConstraint(double min, double max, double rate, long... dimensions){
         this(min, max, rate, Collections.<String>emptySet(), dimensions);
     }
 
@@ -81,7 +81,7 @@ public class MinMaxNormConstraint extends BaseConstraint {
      *                       be dimension 1. For CNNs, this should be dimensions [1,2,3] corresponding to last 3 of
      *                       parameters which have order [depthOut, depthIn, kH, kW]
      */
-    public MinMaxNormConstraint(double min, double max, double rate, Set<String> paramNames, int... dimensions){
+    public MinMaxNormConstraint(double min, double max, double rate, Set<String> paramNames, long... dimensions){
         super(paramNames, dimensions);
         if(rate <= 0 || rate > 1.0){
             throw new IllegalStateException("Invalid rate: must be in interval (0,1]: got " + rate);
