+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/BaseBroadcastBoolOp.java
@@ -44,13 +44,13 @@ import java.util.Map;
 @NoArgsConstructor
 @Slf4j
 public abstract class BaseBroadcastBoolOp extends BaseOp implements BroadcastOp {
-    protected int[] dimension;
+    protected long[] dimension;
 
 
     public BaseBroadcastBoolOp(SameDiff sameDiff,
                                SDVariable i_v1,
                                SDVariable i_v2,
-                               int[] dimension) {
+                               long[] dimension) {
         this(sameDiff, i_v1, i_v2, false, dimension);
     }
 
@@ -58,7 +58,7 @@ public abstract class BaseBroadcastBoolOp extends BaseOp implements BroadcastOp 
                                SDVariable i_v1,
                                SDVariable i_v2,
                                boolean inPlace,
-                               int[] dimension) {
+                               long[] dimension) {
         super(sameDiff, inPlace, new Object[]{i_v2});
         if (i_v1 != null && i_v2 != null) {
             this.sameDiff = sameDiff;
@@ -78,7 +78,7 @@ public abstract class BaseBroadcastBoolOp extends BaseOp implements BroadcastOp 
     public BaseBroadcastBoolOp(SameDiff sameDiff,
                                SDVariable i_v1,
                                SDVariable i_v2,
-                               int[] dimension,
+                               long[] dimension,
                                Object[] extraArgs) {
         super(sameDiff, extraArgs);
         this.dimension = dimension;
@@ -94,7 +94,7 @@ public abstract class BaseBroadcastBoolOp extends BaseOp implements BroadcastOp 
     }
 
 
-    public BaseBroadcastBoolOp(SameDiff sameDiff, SDVariable i_v, int[] dimension, boolean inPlace) {
+    public BaseBroadcastBoolOp(SameDiff sameDiff, SDVariable i_v, long[] dimension, boolean inPlace) {
         this(sameDiff, i_v, i_v.getShape(), inPlace, dimension, null);
     }
 
@@ -102,7 +102,7 @@ public abstract class BaseBroadcastBoolOp extends BaseOp implements BroadcastOp 
                                SDVariable i_v,
                                long[] shape,
                                boolean inPlace,
-                               int[] dimension,
+                               long[] dimension,
                                Object[] extraArgs) {
         super(sameDiff, inPlace, extraArgs);
         this.dimension = dimension;
@@ -121,12 +121,12 @@ public abstract class BaseBroadcastBoolOp extends BaseOp implements BroadcastOp 
 
     public BaseBroadcastBoolOp(SameDiff sameDiff,
                                SDVariable i_v,
-                               int[] dimension,
+                               long[] dimension,
                                Object[] extraArgs) {
         this(sameDiff, i_v, i_v.getShape(), false, dimension, extraArgs);
     }
 
-    public BaseBroadcastBoolOp(INDArray x, INDArray y, INDArray z, int... dimension) {
+    public BaseBroadcastBoolOp(INDArray x, INDArray y, INDArray z, long... dimension) {
         super(x, y, z);
         Broadcast.validateBroadcastDims(x,y,z, dimension);
 
@@ -160,7 +160,7 @@ public abstract class BaseBroadcastBoolOp extends BaseOp implements BroadcastOp 
 
 
     @Override
-    public int[] getDimension() {
+    public long[] getDimension() {
         if (dimension == null) {
             if(x != null && y != null){
                 dimension = Shape.getBroadcastDimensions(x.shape(), y.shape());
@@ -173,7 +173,7 @@ public abstract class BaseBroadcastBoolOp extends BaseOp implements BroadcastOp 
 
 
     @Override
-    public void setDimension(int... dimension) {
+    public void setDimension(long... dimension) {
         this.dimension = dimension;
     }
